Adds a button the user can use to view the location of an item
Desired functionality of displaying a map within the app isn't possible with the `react-native-open-map` library unfortunately.

Currently there is a button within `PostingDetailScreen.js` that opens the devices native map app and displays the location of the item (if the item has point data, otherwise it just does a console log).

This could still be useful, could add/subract a small value from each coordinate value to display the general area of an item to the user. I think we'd need to get a google maps API key to display a map within the app, and it's a significantly more involved process.